Move cache expiration handling to StatsInsightsCache
